requirements.txt
PyYAML

#package
validators
python-gnupg
requests
pycryptodome
tqdm
